Add vanished metadata value
This will do what asked in Feature Request #3121

Please check if I called the plugin correctly as it is the only thing im not 100% sure of, if that is right then it works perfectly! 

What Im asking if it is correct is this: FixedMetadataValue(Essentials, true) , the value should be FixedMetadataValue(Plugin, boolean)

Thanks!